touchpad: fix stuck finger after a click
On a touchpad without resolution, the pinned finger was stuck. The motion
distance scale factor ended up as 0 and the finger never reached the threshold
of 3mm.
int was not the best choice of datatype for a value of 0.007...

Fix the data types for xdist/ydist at the same time, clamping to int may cause
erroneous (un)pinning.

Introduced in 8025b374d564f4a30b089e5cf6fd65e0c6af8da2

https://bugs.freedesktop.org/show_bug.cgi?id=91070

+++ src/evdev-mt-touchpad-buttons.c
@@ -741,7 +741,7 @@ tp_init_buttons(struct tp_dispatch *tp,
 	/* pinned-finger motion threshold, see tp_unpin_finger.
 	   The MAGIC for resolution-less touchpads ends up as 2% of the diagonal */
 	if (device->abs.fake_resolution) {
-		const int BUTTON_MOTION_MAGIC = 0.007;
+		const double BUTTON_MOTION_MAGIC = 0.007;
 		width = abs(absinfo_x->maximum - absinfo_x->minimum);
 		height = abs(absinfo_y->maximum - absinfo_y->minimum);
 		diagonal = sqrt(width*width + height*height);
